changed internal variables/calculations to int64_t fixed precision (#34)
* changed internal variables/calculations to int64_t fixed precision from double.
-currently 4 decimal points used for precision (could be changed by changing types.h > PRICE_PRECISION)

+++ include/backtest-cpp/performance.h
@@ -6,7 +6,7 @@
 
 struct EquityPoint {
     int64_t time;
-    double equity;
+    int64_t equity;
 };
 
 enum class Frequency { DAILY, MINUTE, HOURLY };
